+++ src/components/organisms/CreateScheduleModal.tsx
@@ -15,12 +15,7 @@ type Props = {
   setScheduleList: (scheduleList: Schedule[]) => void;
 };
 
-if (typeof window !== "undefined") {
-const app = document.createElement("div");
-app.id = "root";
-  document.body.appendChild(app);
-  Modal.setAppElement("#root");
-}
+
 
 export const CreateScheduleModal = ({
   isOpen,
